Only delete amarok-appletsrc with KDE 4.5.0.
I've added a runtime version check for this workaround, hoping that this
Plasma bug will be fixed in 4.5.1.

+++ src/context/ContextScene.cpp
@@ -43,11 +43,13 @@ ContextScene::~ContextScene()
 
 void ContextScene::loadDefaultSetup()
 {
+    // WORKAROUND for bug in KDE 4.5.0:
     // Delete amarok-appletsrc config file (created by Plasma), because Plasma tries
-    // to load all applets listed in there, which can lead to duplicated applets which are
-    // not correctly initialized, and all sorts of crashes.
+    // to load all applets listed in there, which can lead to crashes due to applets
+    // being loaded twice.
     // See: BUG 246756
-    QFile::remove( KStandardDirs::locateLocal( "config", "amarok-appletsrc", false ) );
+    if( KDE::versionMajor() == 4 && KDE::versionMinor() == 5 && KDE::versionRelease() == 0 )
+        QFile::remove( KStandardDirs::locateLocal( "config", "amarok-appletsrc", false ) );
 
     Plasma::Containment* c = addContainment( "amarok_containment_vertical" );
     c->setScreen( -1 );
